Fix submenu navigation in NavigationViewItem
The feature for submenus in NavigationViewTop was previously implemented
but left unused prolly because of odd behavior such as stack overflow errors in
AutoSuggest.

The problem was traced to the NavigationViewItem.MenuItems dependency
property, where a new ObservableCollection was incorrectly used as the default
value. According to [MS Docs], "dependency property metadata
shouldn't include a default reference-type value because that value will
be assigned to all instances of the class, creating a singleton class."