Eliminate un-selectable whitespace by ensuring there's no margin between ComboBoxItems
Currently, `ComboBoxItem` and `Border` both have margins equal to `ComboBoxItemMargin`. This gives an effective margin of `2 * ComboBoxItemMargin`.

By removing the `ComboBoxItem` margin and setting `Border`'s margin to `2 * ComboBoxItemMargin` we ensure there's no space between `ComboBoxItem`s that can't be clicked to change the selection.

A downside of this is that the customizability of `ComboBoxItem` is diminished. It is possible to increase the margin but not decrease it under `2 * ComboBoxItemMargin`