Fixed a couple of things here: 1. FonIcon/SymbolIcon markup now fully supports fontsize. Setting the fontsize in the markup will be prioritized over inheriance.
2. Using Content="Text" in the UiButton class can now format the string from inheritance.
Right now it is not possible to do this for the Button class as it breaks the direct string rendering.
Formating text in seperate child elements such as TextBlock does not inheritance as it is not possible to get both working right now.
So better we use this method to format strings because we can alwasys format any TextBlock on it's own inside the UiButton.

3. Moved IconAlignment into IconElement folder and added VerticalIconAlignment as possibility.
Docs added (check them, maybe you don't like them)

4. Added a border to the Slider style for visual value feedback.

5. Refactored the SystemThemeManager class.
Using SystemUsesLightTheme is unnecessary since the app theme is the only one necessary variable we need.
If the system uses light theme withouth seperating the app theme, SystemUsesLightTheme will always return 1 and vice versa.
We should also take into account if people change their app theme separately and base it on that one.